refactor: Move helper to transactions
The trigger for this move was that categories helpers have a dependency
chain that at some point import cozy-ui, and it makes it difficult to
use with node js CLI. Anyway the helper was better suited to be in
transaction helpers.
