+++ src/ducks/transactions/TransactionCategoryEditor.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useClient } from 'cozy-client'
 import { updateTransactionCategory } from 'ducks/transactions/helpers'
-import { getCategoryId } from 'ducks/categories/helpers'
+import { getCategoryId } from 'ducks/transactions/helpers'
 import CategoryChoice from 'ducks/categories/CategoryChoice'
 
 /**
